# Wavelet/docs/swagger.yaml
basePath: /
definitions:
...
  pages.DeploymentFileView:
    properties:
      checksum:
        type: string
      created_at:
        type: string
      deployment_id:
        type: integer
      id:
        type: integer
      path:
        type: string
      size:
        type: integer
    type: object
  pages.DeploymentView:
    properties:
      activated_at:
        type: string
      checksum:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      deployment_number:
        type: integer
      file_count:
        type: integer
      id:
        type: integer
      project_id:
        type: integer
      status:
        type: string
      total_size:
        type: integer
    type: object
  pages.Input:
    properties:
      api_proxy_enabled:
        type: boolean
      api_proxy_pass:
        type: string
      api_proxy_path:
        type: string
      api_proxy_rewrite:
        type: string
      description:
        type: string
      enabled:
        type: boolean
      entry_file:
        type: string
      name:
        type: string
      root_dir:
        type: string
      slug:
        type: string
      spa_fallback_enabled:
        type: boolean
      spa_fallback_path:
        type: string
    type: object
  pages.View:
    properties:
      active_deployment:
        $ref: '#/definitions/pages.DeploymentView'
      active_deployment_id:
        type: integer
      api_proxy_enabled:
        type: boolean
      api_proxy_pass:
        type: string
      api_proxy_path:
        type: string
      api_proxy_rewrite:
        type: string
      created_at:
        type: string
      deployment_count:
        type: integer
      description:
        type: string
      enabled:
        type: boolean
      entry_file:
        type: string
      id:
        type: integer
      name:
        type: string
      root_dir:
        type: string
      slug:
        type: string
      spa_fallback_enabled:
        type: boolean
      spa_fallback_path:
        type: string
      updated_at:
        type: string
    type: object
...
  response.Any:
    properties:
      data: {}
      error_msg:
        example: ""
        type: string
    type: object
...
info:
  contact:
    name: OpenFlare
    url: https://github.com/Rain-kl/OpenFlare
  description: OpenFlare 平台后端 API，提供用户认证、系统配置、任务调度与边缘节点管理能力。
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: OpenFlare API
  version: 1.0.0
paths:
...
  /api/v1/openflare/pages:
    get:
      description: 返回全部 OpenFlare Pages 项目，需要管理员权限
      produces:
      - application/json
      responses:
        "200":
          description: Pages 项目列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/pages.View'
                  type: array
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Any'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/response.Any'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 列出 Pages 项目
      tags:
      - openflare-pages
    post:
      consumes:
      - application/json
      description: 创建新的 OpenFlare Pages 项目，需要管理员权限
      parameters:
      - description: 项目参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pages.Input'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功的项目
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  $ref: '#/definitions/pages.View'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Any'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/response.Any'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 创建 Pages 项目
      tags:
      - openflare-pages
  /api/v1/openflare/pages/{id}:
    get:
      description: 按 ID 返回 Pages 项目详情，需要管理员权限
      parameters:
      - description: 项目 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Pages 项目详情
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  $ref: '#/definitions/pages.View'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Any'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/response.Any'
        "404":
          description: 项目不存在
          schema:
            $ref: '#/definitions/response.Any'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 获取 Pages 项目详情
      tags:
      - openflare-pages
  /api/v1/openflare/pages/{id}/delete:
    post:
      description: 按 ID 删除 OpenFlare Pages 项目，需要管理员权限
      parameters:
      - description: 项目 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/response.Any'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Any'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/response.Any'
        "404":
          description: 项目不存在
          schema:
            $ref: '#/definitions/response.Any'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 删除 Pages 项目
      tags:
      - openflare-pages
  /api/v1/openflare/pages/{id}/deployments:
    get:
      description: 返回指定项目的全部部署记录，需要管理员权限
      parameters:
      - description: 项目 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 部署列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/pages.DeploymentView'
                  type: array
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Any'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/response.Any'
        "404":
          description: 项目不存在
          schema:
            $ref: '#/definitions/response.Any'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 列出 Pages 部署
      tags:
      - openflare-pages
  /api/v1/openflare/pages/{id}/deployments/{deployment_id}/activate:
    post:
      description: 将指定部署设为项目当前生效版本，需要管理员权限
      parameters:
      - description: 项目 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 部署 ID
        in: path
        name: deployment_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 激活后的项目
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  $ref: '#/definitions/pages.View'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Any'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/response.Any'
        "404":
          description: 项目或部署不存在
          schema:
            $ref: '#/definitions/response.Any'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 激活 Pages 部署
      tags:
      - openflare-pages
  /api/v1/openflare/pages/{id}/deployments/{deployment_id}/delete:
    post:
      description: 删除指定项目的部署记录，需要管理员权限
      parameters:
      - description: 项目 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 部署 ID
        in: path
        name: deployment_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/response.Any'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Any'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/response.Any'
        "404":
          description: 项目或部署不存在
          schema:
            $ref: '#/definitions/response.Any'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 删除 Pages 部署
      tags:
      - openflare-pages
  /api/v1/openflare/pages/{id}/deployments/upload:
    post:
      consumes:
      - multipart/form-data
      description: 为指定项目上传 ZIP 部署包，需要管理员权限
      parameters:
      - description: 项目 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 部署包 ZIP 文件
        in: formData
        name: package
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 部署记录
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  $ref: '#/definitions/pages.DeploymentView'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Any'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/response.Any'
        "404":
          description: 项目不存在
          schema:
            $ref: '#/definitions/response.Any'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 上传 Pages 部署包
      tags:
      - openflare-pages
  /api/v1/openflare/pages/{id}/update:
    post:
      consumes:
      - application/json
      description: 按 ID 更新 OpenFlare Pages 项目，需要管理员权限
      parameters:
      - description: 项目 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 项目参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pages.Input'
      produces:
      - application/json
      responses:
        "200":
          description: 更新后的项目
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  $ref: '#/definitions/pages.View'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Any'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/response.Any'
        "404":
          description: 项目不存在
          schema:
            $ref: '#/definitions/response.Any'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 更新 Pages 项目
      tags:
      - openflare-pages
  /api/v1/openflare/pages/deployments/{deployment_id}/files:
    get:
      description: 返回指定部署包含的文件清单，需要管理员权限
      parameters:
      - description: 部署 ID
        in: path
        name: deployment_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 部署文件列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/pages.DeploymentFileView'
                  type: array
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Any'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/response.Any'
        "404":
          description: 部署不存在
          schema:
            $ref: '#/definitions/response.Any'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 列出 Pages 部署文件
      tags:
      - openflare-pages
...
securityDefinitions:
  SessionCookie:
    in: cookie
    name: session
    type: apiKey
swagger: "2.0"